requirements.txt
streamlit
tensorflow
Pillow